Detangled recipe service from the media service when handling GET, POST, PUT, and DELETE requests for recipe media.

+++ app/services/media.js
@@ -80,7 +80,7 @@ const writeToDisk = async (id, files) => {
 const set = async (id, files) => {
     try {
         // check if no media was given to upload
-        if (!files) {
+        if (files.cleared.length === 0 && files.rejected.length === 0) {
             const message = 'No media to upload for entity with ID of' +
                             ` "${id}", nothing to do.`
             return quickResponse(204, message)
@@ -149,8 +149,38 @@ const unset = async id => {
  */
 const reset = async (id, files) => {
     try {
-        await removeDir(id)
-        return await set(id, files)
+        // check if no media was given to upload
+        if (files.cleared.length === 0 && files.rejected.length === 0) {
+            const message = 'No media to upload for entity with ID of' +
+                            ` "${id}", nothing to do.`
+            return quickResponse(204, message)
+        }
+        const { cleared, rejected, filteredFiles } = files
+
+        // default control variables based on success
+        let message = 'Some of the selected media was unable to be uploaded.'
+        let status = 201
+        let writeMedia = true
+
+        const nothingWasCleared = cleared.length === 0
+        const nothingWasRejected = rejected.length === 0
+        if (nothingWasCleared) {
+            message = 'The selected media was unable to be uploaded,' +
+                      ' nothing to do.'
+            status = 204
+            writeMedia = false
+        } else if (nothingWasRejected) {
+            message = 'The media for entity with ID of' +
+                      ` "${id}" was successfully uploaded.`
+        }
+
+        // don't write media if there is nothing to upload
+        if (writeMedia) {
+            await removeDir(id)
+            await createDir(id)
+            await writeToDisk(id, filteredFiles)
+        }
+        return quickResponse(status, message, { cleared, rejected })
     } catch (err) {
         return quickResponse(500)
     }
